+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Add `dx-data-grid-generic` storybook story
 - Add `icon` property to the `dx-chip` component
+- Added a new storybook for the theme inspector and also added some new color tokens.
 
 ### Fixed
 - Fix the `XS` size of the `dx-dialog` component